Reformating of classes

+++ HW5/Running/Member.java
@@ -2,11 +2,11 @@ package HW5.Running;
 
     abstract class Member{
 // якщо не буде доступа к макс длине то треба зробити геттери (в батьківському класі стоїть протектед!
-        protected double maxLenght;
-        protected double maxHight;
+        protected int maxLenght;
+        protected int maxHight;
 
-        public double getMaxLenght() {return maxLenght;}
-        public double getMaxHight() {return  maxHight;}
+        public int getMaxLenght() {return maxLenght;}
+        public int getMaxHight() {return  maxHight;}
 
         abstract void run();
         abstract void jump();
